refactor(session-store): extract run projection owner
